optimize memory by using bitfields for bool and smaller int types
This commit introduces several memory optimizations across various UDisks components.
`gboolean` flags in `UDisksLinuxProviderPrivate`, `UDisksConfigManager`, and `UDisksLinuxBlockPrivate` have been converted to 1-bit bit-fields. This change packs multiple boolean flags into a single byte or word, significantly reducing the memory footprint of instances of these structures.
Additionally, the `freq` and `passno` members in `UDisksFstabEntry` have been changed from `gint` to `gint16`. Since the expected values for these fields are typically small, using a 16-bit integer type instead of a 32-bit one further contributes to memory savings for each fstab entry.
These optimizations aim to reduce overall memory consumption, which is beneficial for system-level services like udisks.

References:
*   [GeeksforGeeks: Bit Fields in C/C++](https://www.geeksforgeeks.org/bit-fields-c/)
*   [cppreference.com: Bit field](https://en.cppreference.com/w/cpp/language/bit_field)
*   [GeeksforGeeks: Structure Member Alignment, Padding and Data Packing](https://www.geeksforgeeks.org/structure-member-alignment-padding-and-data-packing/)
*   [IBM Knowledge Center: C data types (explains sizes)](https://www.ibm.com/docs/en/ztpf/2019?topic=types-c-data-types)
*   [Embedded.com: Data structure alignment and packing](https://www.embedded.com/data-structure-alignment-and-packing/)

+++ src/udisksfstabentry.c
@@ -50,8 +50,8 @@ struct _UDisksFstabEntry
   gchar *dir;
   gchar *type;
   gchar *opts;
-  gint freq;
-  gint passno;
+  gint16 freq;
+  gint16 passno;
 };
 
 typedef struct _UDisksFstabEntryClass UDisksFstabEntryClass;
